resume_extraction/input_processing/docx_parser.py
import docx

def extract_text_from_docx(file_path):

    doc = docx.Document(file_path)

    text = []

    for para in doc.paragraphs:
        text.append(para.text)

    return "\n".join(text)