resume_extraction/requirements.txt
spacy
pymupdf
pytesseract
pillow
python-docx
gdown